+++ acquisition/micromotion_tracking.py
@@ -2,503 +2,379 @@ import os
 import cv2
 import numpy as np
 import math
+from tracking_methods import get_frame, set_up_detector, setup_tracking
 
-# ----------------- Usage Instructions ----------------- #
-
-'''
-First upload the video file of choice and run it in "binary" viewtype to find the object and center the frame on it
-with the xrange and yrange values. The surface of the trap must be set to the bottom of the visible frame for 
-accurate height data output. A bottom_bar value is likely necessary to cut out the noise from the lowest n frames
-Then, set the "auto" feature to True, and the code will run and output data automatically.
-If particle tracking is sparse, set "poortracking" to True and input the particle indices to "index_list". Indices can
-be identified by manually running the code with auto = False and observing the green value near the object's centroid
-'''
-
-# --------------------------- Parameter Definition ---------------------------------------------- #
-
-
-videofile = cv2.VideoCapture('acquisition/8-16Trial4.avi')   # Specifies the video file
-viewtype = "image"   # "binary" for binary image, "image" for annotated video, "frame" for original video, "cleanthresh" for undialated
-output = "none"   # "tuple" for tuple output, "none" for none
-showquant = "none"   # height, micromotion, or both can be displayed on the image
-auto = "False"   # Automatically Runs Code (will not display video)
-poortracking = False   # Will ignore index of the particle if varying. ONLY USE IF NO OTHER PARTICLES IN FRAME
-fps = 20   # Frame rate of the camera
-changeinterval = 5   # How many seconds between data point collection
-points_to_image = []    # Saves an image of the frame at the specified data point (3, 20, 31, 33, 35 for paper figure 3(a))
-sample_frames = 15   # How many frames the data is averaged over
-binthresh = 26   # Binary threshold distinguishing white from black (higher for brighter pixels, lower for darker)
-xrange = (800, 1200)   # define the visible x range (images are usually 1616 x 1240)
-yrange = (547, 933)   # define the visible y range
-bottom_bar = 100   # sets the height of bottom blocker. Used to block unwanted noise within the visible frame
-top_bar = 0   # sets the height of top blocker
-left_bar = 0   # sets the height of left blocker
-right_bar = 0   # # sets the height of right blocker
-
-indexlist = []   # Specifies the multiple indices of desired tracking particle in case of sparse tracking/regular index shifts
+class TrackingConfig:
+    def __init__(self):
+        self.video_file = "acquisition/ExampleMicromotion.avi"
+        self.view_type = "image"        # "image" to block out white binary noise, "binary" to block out black binary noise
+        self.start_frame = 0            # Defines starting frame. ONLY FOR DEBUGGING
+        self.fps = 20                   # fps of the camera
+        self.start_voltage = 40         # Initial voltage value 
+        self.voltage_increment = 5      # Voltage step between datapoints
+        self.change_interval = 5        # Time between data points in the real-time trial (seconds)
+        self.sample_frames = 15         # Number of frames averaged over per data point
+        self.bin_thresh = 26            # Binary threshold for object detection
+        self.x_range = (200, 900)       # x-axis frame of interest limits
+        self.y_range = (554, 1000)      # y-axis frame of interest limits
+        self.bottom_bar = 100           # Erasure rectangle, measured in pixels from the bottom edge
+        self.top_bar = 0                # Erasure rectangle, measured in pixels from the top edge
+        self.left_bar = 0               # Erasure rectangle, measured in pixels from the left edge
+        self.right_bar = 0              # Erasure rectangle, measured in pixels from the right edge
+        self.pixel_to_mm = 0.01628      # Pixel-to-millimeter conversion, gathered from calibration image. "None" will output raw pixel data
 
 
-#-------------------------------Initializing parameter variations-------------------------------------- #
+def get_default_config():
+    return TrackingConfig()
 
 
-ticker = 0
-val = 0
-
-# Sets which frames the data collection starts and ends at. The offset is set to 40% of the spacing between points to ensure equilibrium
-offset = (fps * changeinterval) * 0.4
-collection_frames = []
-for i in range(100):
-    collection_frames.append(int((fps * changeinterval * i) + offset))
-end_collection_frames = []
-for i in range(100):
-    end_collection_frames.append(int((collection_frames[i] + sample_frames)))
-
-# Adjusts for inverse y-axis indexing
-imageheight = 1240
-x_start, x_end = xrange
-y_start, y_end = (imageheight-yrange[1]), (imageheight-yrange[0])
-
-# Sets the binary rectangles (blockers)
-ylength = y_end - y_start
-xlength = x_end - x_start
-top_rect_pt1, top_rect_pt2 = (0, 0), (1616, top_bar)
-left_rect_pt1, left_rect_pt2 = (0, 0), (left_bar, 1240)
-right_rect_pt1, right_rect_pt2 = (xlength-right_bar, 0), (xlength, 1240)
-bottom_rect_pt1, bottom_rect_pt2 = (0, ylength-bottom_bar), (1616, ylength)
-frameheight = yrange[1]-yrange[0]
-if viewtype == "binary":
-    rectangle_color = (255, 255, 255)
-else: rectangle_color = (0, 0, 0)
-
-# Establishes useful per-datapoint sets for averaging data
-microvec = []
-heightvec = []
-bothvec = []
-datapointnum = 0
-
-print("Running program...")
+def frame_dimensions(cap, frame_num, config = get_default_config()):
+    """
+    Calculate frame dimensions and ranges
+    :param cap: Video capture object from the OpenCV package
+    :param frame_num: Frame number of interest
+    :return x_start, x_end,...: Define the rectangular region of interest
+    """
+    ret, initial_frame = get_frame(cap, frame_num)
+    initial_frame_dim = initial_frame.shape
+    imageheight = initial_frame_dim[0]
+    x_start, x_end = config.x_range
+    y_start, y_end = (imageheight - config.y_range[1]), (imageheight - config.y_range[0])
+    return x_start, x_end, y_start, y_end
 
 
-# --------------------------------------- Setting up the detector ------------------------------------------- #
+def gen_initial_frame(cap, config = get_default_config()):
+    """
+    Generate and display initial frame
+    :param cap: Video capture object from the OpenCV package
+    :return x_start, x_end,...: Define the rectangular region of interest
+    """
+    frame_num = config.start_frame
+    x_start, x_end, y_start, y_end = frame_dimensions(cap, frame_num)
+    ret, initial_frame = get_frame(cap, config.start_frame)
+    cv2.imshow("Frame", initial_frame[y_start:y_end, x_start:x_end])
+    return x_start, x_end, y_start, y_end
 
 
-params = cv2.SimpleBlobDetector.Params()
-
-# Filter by color
-params.filterByColor = True
-params.blobColor = 255
-
-# Filter by area (pixels)
-params.filterByArea = True
-params.minArea = 5
-params.maxArea = 300
-
-# Filter by circularity
-params.filterByCircularity = False
-
-# Filter by convexity
-params.filterByConvexity = False
-
-# Filter by inertia ratio (To detect elongated shapes)
-params.filterByInertia = False
-params.minInertiaRatio = 0.01
-params.maxInertiaRatio = 0.3
-
-# Create a detector with the parameters
-ver = (cv2.__version__).split('.')
-if int(ver[0]) < 3:
-    detector = cv2.SimpleBlobDetector()
-else:
-    detector = cv2.SimpleBlobDetector_create(params)
+def define_blockers(cap, frame_num, config = get_default_config()):
+    """
+    Define blocking rectangles for frame processing
+    :param cap: Video capture object from the OpenCV package
+    :param frame_num: Frame number of interest
+    :return: Tuple object containing tuple elements that define the locations of rectangles for omission
+    """
+    x_start, x_end, y_start, y_end = frame_dimensions(cap, frame_num)
+    ylength = y_end - y_start
+    xlength = x_end - x_start
+    
+    top_rect = ((0, 0), (1616, config.top_bar))
+    left_rect = ((0, 0), (config.left_bar, 1240))
+    right_rect = ((xlength - config.right_bar, 0), (xlength, 1240))
+    bottom_rect = ((0, ylength - config.bottom_bar), (1616, ylength))
+    
+    return (*top_rect, *left_rect, *right_rect, *bottom_rect)
 
 
-# ---------------------------------- Setting program parameters --------------------------------------------- #
+def post_processing(cap, frame, frame_num, config = get_default_config()):
+    """
+    Process frame and apply filters
+    :param cap: Video capture object from the OpenCV package
+    :param frame: Image of the frame returned by cap.read()
+    :param frame_num: Frame number of interest
+    :return roi_frame: Image of the frame, cropped to the region of interest
+    :return closing: Binary image of the frame after erasing small imperfections, cropped to the region of interest
+    :return clean_thresh: "Cleaned" image of the frame with small binary imperfections erased
+    :return closing_raw: Binary image of the frame post-erasure without the binary blocker
+    """
+    x_start, x_end, y_start, y_end = frame_dimensions(cap, frame_num)
+    blockers = define_blockers(cap, frame_num)
+    rectangle_color = (255, 255, 255) if config.view_type == "binary" else (0, 0, 0)
+    cleaning_kernel = np.ones((2, 2), np.uint8)
+    filling_kernel = np.ones((2, 2), np.uint8)
+    roi_frame = frame[y_start:y_end, x_start:x_end]
+    gray_frame = cv2.cvtColor(roi_frame, cv2.COLOR_BGR2GRAY)
+    ret, thresh = cv2.threshold(gray_frame, config.bin_thresh, 255, cv2.THRESH_BINARY)
+    clean_thresh = cv2.morphologyEx(thresh, cv2.MORPH_OPEN, cleaning_kernel, iterations=1)
+    closing_raw = cv2.morphologyEx(thresh, cv2.MORPH_OPEN, cleaning_kernel, iterations=1)
+    for i in range(0, len(blockers), 2):
+        cv2.rectangle(clean_thresh, blockers[i], blockers[i+1], rectangle_color, -1)
+    closing = cv2.morphologyEx(clean_thresh, cv2.MORPH_CLOSE, filling_kernel, iterations=2)
+
+    return roi_frame, closing, clean_thresh, closing_raw
 
 
-# Choosing the video to track in
-vid_cap = videofile
-if not vid_cap.isOpened():
-    print("Error: Could not open video file.")
-    exit()
-ret, start_frame = vid_cap.read()
-print(ret)
-
-# Object index of interest (Selecting a particle to collect data on)
-index_of_interest = 0
-
-# Defining up morphological transformations kernels (Can be different for some optimizations)
-cleaning_kernel = np.ones((2, 2), np.uint8)
-filling_kernel = np.ones((2, 2), np.uint8)
-
-# Define a starting frame number
-start_frame_num = 1
-
-
-# ------------------------------------------ Priming the main loop --------------------------------------------------#
-
-
-# Setting up memory between frames for objects
-keypoints_prev_frame = []
-tracking_objects = {}
-track_id = 0
-
-# Frame counter
-frame_num = start_frame_num
-total_frames = int(vid_cap.get(cv2.CAP_PROP_FRAME_COUNT))
-print(total_frames)
-
-# Trial counter
-trial_num = 1
-first_point = False
-
-# Starting point (For clean graphing)
-start_x = 0
-first_detect = False
-
-# Function to read a specific frame
-def get_frame(cap, got_frame_num):
-    cap.set(cv2.CAP_PROP_POS_FRAMES, got_frame_num)
-    _, got_frame = cap.read()
-    return _, got_frame
-
-
-# ------------------------------------------------- Main Loop ------------------------------------------- #
-
-
-ret, start_frame = get_frame(vid_cap, frame_num)
-
-start_frame_dim = start_frame.shape
-start_frame_height = start_frame_dim[0]
-start_frame_width = start_frame_dim[1]
-print(f"Frame height: {start_frame_height}\n"
-      f"Frame width: {start_frame_width}")
-
-cv2.imshow("Frame", start_frame[y_start:y_end, x_start:x_end])
-
-run = True
-run_body = True
-while run:
-
-    # Initializing the frames for the loop to play
-    frames_to_play = 0
-
-    # Waiting for the esc key to end the loop (DEC: 27 on ASCII table)
-    # Space plays 20 frames (DEC: 32 on ASCII table)
-    # Any other key moves to the next frame
-    key = cv2.waitKey()
-    if key == 27:
-        run = False
-    elif key == 32:
-        index_of_interest = int(input('Enter the index of interest: '))
-        frames_to_play = sample_frames
-
-        first_point = True
+def locate_particles(roi_frame, closing, keypoints_prev_frame, frame_num, tracking_objects, track_id, y_end, y_start, last_known = None, config = get_default_config()):
+    """
+    Locate and track particles in frame
+    :param roi_frame: Image of the frame, cropped to the region of interest
+    :param closing: Binary image of the frame after erasing imperfections, cropped to the region of interest
+    :param keypoints_prev_frame: List object containing tuples of particle locations in the previous frame
+    :param frame_num: Frame number of interest
+    :param tracking_objects: Dictionary object containing particles' locations
+    :param track_id: Index of particle in tracking_objects dictionary
+    :param y_end, y_start: Define the y frame of interest
+    :return x_position: X position of the particle's centroid, indicated in pixels from the left of the frame of interest
+    :return y_position_adj: Y position of the particle's centroid, indicated in pixels from the bottom of the frame of interest
+    :return height: Height of detected object in pixels
+    :return image_with_keypoints: Image of the frame of interest, red circles drawn at the centroid of detected objects
+    :return keypoints_cur_frame: List object containing tuples of particle locations in the current frame
+    """
+    detector = set_up_detector()
+    keypoints = detector.detect(closing)
+    keypoints_cur_frame = []
+    x_position, y_position_adj, height = "NaN", "NaN", "NaN"
+    for keypoint in keypoints:
+        keypoints_cur_frame.append(keypoint.pt)
+    keypoints_copy = keypoints_cur_frame
+    image_with_keypoints = cv2.drawKeypoints(roi_frame, keypoints, np.array([]), (0, 0, 255))
+    contours, _ = cv2.findContours(closing, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+    if frame_num <= config.start_frame + 2:
+        track_id = _initialize_tracking(keypoints_cur_frame, keypoints_prev_frame, tracking_objects, track_id)
     else:
-        if auto == "False":
-            frames_to_play = 1
-            showvid = True
-            datacollect = True
-        if auto == "True":
-            print('Beginning Autoscan...')
-            frames_to_play = 999999
-            showvid = False
-            datacollect = False
-
-
-    # Running through the specified number of frames
-    for i in range(frames_to_play):
-        ret, frame = get_frame(vid_cap, frame_num)
-        if not ret and frame_num < total_frames:
-            print("Cannot retrieve frame.")
-            run = False
-        elif frame_num >= total_frames:
-            sit = True
-            while sit:
-                end_control = cv2.waitKey(5)
-                if end_control == 27:
-                    run_body = False
-                    run = False
-                    sit = False
-
-        # Handle edge case: Running anything after the last frame causes an unnecessary error
-        if not run_body:
-            break
-
-        # Frame information (Useful for debugging)
-        dimensions = frame.shape
-        height = dimensions[0]
-        width = dimensions[1]
-        # print(f"Height: {height}\nWidth: {width}")
-
-        # Creating the region of interest frame
-        roi_frame = frame[y_start:y_end, x_start:x_end]
-        roi_height = roi_frame.shape[0]
-        roi_width = roi_frame.shape[1]
-        # print(f"Height: {roi_height}\nWidth: {roi_width}")
-
-        # Converting to black and white
-        gray_frame = cv2.cvtColor(roi_frame, cv2.COLOR_BGR2GRAY)
-
-        # Apply binary thresholding (Converting to intensity values of 0 and 255)
-        ret, thresh = cv2.threshold(gray_frame, binthresh, 255, cv2.THRESH_BINARY)
-
-        # Post-processing
-        clean_thresh = cv2.morphologyEx(thresh, cv2.MORPH_OPEN, cleaning_kernel, iterations=1)
-        cv2.rectangle(clean_thresh, top_rect_pt1, top_rect_pt2, rectangle_color, -1)  # Top Erasure
-        cv2.rectangle(clean_thresh, left_rect_pt1, left_rect_pt2, rectangle_color, -1)  # Left Erasure
-        cv2.rectangle(clean_thresh, right_rect_pt1, right_rect_pt2, rectangle_color, -1)  # Right Erasure
-        cv2.rectangle(clean_thresh, bottom_rect_pt1, bottom_rect_pt2, rectangle_color, -1)  # Bottom erasure
-        closing = cv2.morphologyEx(clean_thresh, cv2.MORPH_CLOSE, filling_kernel, iterations=2)
-
-        # Finding the locations of the particles (Will return the top left point of the detected object)
-        keypoints = detector.detect(closing)
-
-        keypoints_cur_frame = []
-        for keypoint in keypoints:
-            point = keypoint.pt
-            keypoints_cur_frame.append(keypoint.pt)
-
-        # If the outimage parameter receives an argument of None or an empty array, it will be a copy of the source image
-        image_with_keypoints = cv2.drawKeypoints(roi_frame, keypoints, np.array([]), (0, 0, 255))
-
-        # Finding the contours
-        contours, _ = cv2.findContours(closing, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-
-        # Compare prev and cur point only at the beginning
-        if frame_num <= 2:
-            for pt1 in keypoints_cur_frame:
-                for pt2 in keypoints_prev_frame:
-                    distance = math.dist(pt1, pt2)
-
-                    # Add an identified object to the tracking_objects dict and prepare the next id
-                    if distance < 10:
-                        tracking_objects[track_id] = [pt1]
-                        track_id += 1
-        else:
-            # Create a copy bc we can't remove dict elements while traversing it
-            tracking_objects_copy = tracking_objects.copy()
-            keypoints_cur_frame_copy = keypoints_cur_frame.copy()
-
-            for object_id, item2 in tracking_objects_copy.items():
-                object_exists = False
-                for pt1 in keypoints_cur_frame:
-                    distance = math.dist(pt1, item2[0])
-
-                    # Update object position
-                    if distance < 10:
-                        tracking_objects[object_id] = [pt1]
-                        object_exists = True
-                        if pt1 in keypoints_cur_frame:
-                            keypoints_cur_frame.remove(pt1)
-                        continue  # Moves on to checking next id once we confirm existence
-
-                # Remove lost IDs
-                if not object_exists:
-                    tracking_objects.pop(object_id)
-
-        # Add found IDs
-        for pt1 in keypoints_cur_frame:
-            tracking_objects[track_id] = [pt1]
-            track_id += 1
-            # Iterate over each contour to find the bounding rectangle and get the height
-        for i in contours:
-            # Get the bounding rectangle for each contour (x, y indicate the top left point of the object)
-            x, y, w, h = cv2.boundingRect(i)
-            centroid_x, centroid_y = int(x+w/2), int(y+h/2)
-            adj_centroid_y = (y_end - y_start) - centroid_y
-
-            # Check which tracking object is in the bounding box
-            for key in tracking_objects.keys():
-                if x <= tracking_objects[key][0][0] <= x + w and y <= tracking_objects[key][0][1] <= y + h:
-                    tracking_objects[key].append(h)
-
-            # Optionally, draw the bounding rectangle on the original image
-            if viewtype == "image" and auto == "True":
-                showvid = True
-            else:
-                pass
-            if showvid == True:
-                # Displays specified quantities on the image
-                cv2.rectangle(image_with_keypoints, (x, y), (x + w, y + h), (0, 255, 0), 1)
-                if showquant == "height":
-                    cv2.putText(image_with_keypoints, 'y: ' + str(adj_centroid_y), (centroid_x, centroid_y + h + 15), 0, 0.5, (0, 255, 0), 1)
-                if showquant == "micro":
-                    cv2.putText(image_with_keypoints, 'a: ' + str(h), (centroid_x, centroid_y + h + 15), 0, 0.5, (0, 255, 0), 1)
-                if showquant == "both":
-                    cv2.putText(image_with_keypoints, 'y:' + str(adj_centroid_y), (centroid_x + 12, centroid_y - 8), 0, 0.34, (0, 255, 0), 1)
-                    cv2.putText(image_with_keypoints, 'a:' + str(h), (centroid_x + 12, centroid_y + 8), 0,0.34, (0, 255, 0), 1)
-            else:
-                pass
-            if viewtype == "image" and auto == "True":
-                showvid = False
-            else:
-                pass
-
-        # Storing heights detected in "Space bar" frames
-        try:
-            x, y = point
-            adj_y = (y_end - y_start) - y
-        except:
-            NameError
-        if datacollect == True:
+        _update_tracking(keypoints_cur_frame, tracking_objects)
+    _process_contours(contours, tracking_objects)
+    if frame_num >= (config.start_frame + 2) and len(tracking_objects.keys()) > 0:
+        tracking_objects_copy = tracking_objects.copy()
+        for i in tracking_objects_copy.keys():
             try:
-                x, y = point
-                adj_y = (y_end - y_start) - y
-            except NameError:
-                pass
-        if poortracking == True:
-            for i in range(1000):
-                try:
-                    height_of_interest = tracking_objects[i][1]
-                    if len(tracking_objects.keys()) > 0 and height_of_interest > 0:
-                        if i in indexlist:
-                            try:
-                                heightvec.append(adj_centroid_y)
-                                microvec.append(height_of_interest)
-                            except KeyError:
-                                pass
-                            break
-                        else:
-                            pass
-                except KeyError or IndexError:
-                    pass
-        if poortracking == False:
-            try:
-                height_of_interest = tracking_objects[index_of_interest][1]  # height_of_interest = tracking_objects[index_of_interest][1]
-                if len(tracking_objects.keys()) > 0:
-                    try:
-                        heightvec.append(adj_centroid_y)
-                        microvec.append(height_of_interest)
-                    except KeyError:
-                        pass
-            except KeyError or IndexError:
-                pass
-
-        # Puts frame labels and displays requested viewtype frame
-        if showvid == True:
-            for object_id, item in tracking_objects.items():
-                cv2.putText(image_with_keypoints, str(object_id), (int(item[0][0] - 5), int(item[0][1] - 17)), 0, 0.5,
-                            (0, 255, 0), 1)
-            # Showing the current frame annotated with the keypoints
-            if frame_num >= total_frames - 1:
-                cv2.putText(image_with_keypoints, "Frame: end", (5, 20), 0, 0.5, (0, 255, 0), 1)
-            else:
-                cv2.putText(image_with_keypoints, "Frame: " + str(frame_num), (5, 20), 0, 0.5, (0, 255, 0), 1)
-            if viewtype == "binary":
-                cv2.imshow("Frame", closing) #image_with_keypoints, closing
-            if viewtype == "image":
-                cv2.imshow("Frame", image_with_keypoints)
-            if viewtype == "frame":
-                cv2.imshow("Frame", frame)
-
-            cv2.waitKey(50)
-
-        # Begins data collection if the frame number is at one of interest
-        else:
-            if frame_num in collection_frames:
-                datacollect = True
-            if frame_num in end_collection_frames:
-                ticker = ticker + 1
-                print(ticker)
-
-                # Identifies if frame is one to save an image from and takes a zoomed in image of the particle
-                if ticker in points_to_image:
-                    val = val + 1
-                    if viewtype == "binary":
-                        x, y = point
-                        x, y = int(x), int(y)
-                        cropped_image = closing[y - 40:y + 40, x - 50:x + 50]
-                        x, y = point
-                        x, y = int(x), int(y)
-                        if val == 1:
-                            cv2.imshow("Image", cropped_image)
-                            cv2.waitKey(0)
-                            cv2.destroyAllWindows()
-                            cropped_image = cv2.cvtColor(cropped_image, cv2.COLOR_GRAY2BGR)
-                            cv2.imshow('enlarged_image', cropped_image)
-                            cv2.imwrite("Frame" + str(points_to_image[val - 1]) + str(viewtype) + "Stacked.tif",
-                                        cropped_image)
-                            cv2.waitKey(0)
-                            cv2.destroyAllWindows()
-                        else:
-                            cropped_image = closing[y - 40:y + 40, x - 50:x + 50]
-                            cv2.imwrite("Frame" + str(points_to_image[val - 1]) + str(viewtype) + "Stacked.tif",
-                                        cropped_image)
-                    if viewtype == "image":
-                        x, y = point
-                        x, y = int(x), int(y)
-                        cropped_image = image_with_keypoints[y - 50:y + 40, x - 50:x + 50]
-                        if val == 1:
-                            cv2.imshow("Image", cropped_image)
-                            cv2.waitKey(0)
-                            cv2.destroyAllWindows()
-                        else:
-                            pass
-                        cv2.imwrite("Frame" + str(points_to_image[val - 1]) + str(viewtype) + "Stacked.tif",
-                                    cropped_image)
-                    if viewtype == "frame":
-                        framed_image = frame[y_start:y_end, x_start:x_end]
-                        x,y = point
-                        x, y = int(x), int(y)
-                        print(x,y)
-                        cropped_image = framed_image[y-50:y+50, x-50:x+70]
-                        cv2.imwrite("Frame" + str(points_to_image[val - 1]) + str(viewtype) + "Final.tif", cropped_image)
-                    print('Click!')
-
-                datapointnum = datapointnum + 1
-
-                # Stores the data in the "Tuple.txt" file and prompts if Tuple.txt already contains data upon beginning collection
-                if output == "tuple":
-                    avgheight = round(np.mean(heightvec), 2)
-                    avgmicro = round(np.mean(microvec), 2)
-                    percentage = (frame_num / total_frames) * 100
-                    print(str(round(percentage, 0)) + '% , Average Height = ' + str(avgheight) + ' , Average Micromotion = ' + str(avgmicro) + '\n')
-                    if os.stat('Tuple.txt').st_size != 0:
-                        if datapointnum == 1:
-                            acknowledgement = ""
-                            while acknowledgement != "continue":
-                                acknowledgement = input('Tuple.txt already contains data. Type "continue" to add to the existing file, otherwise stop. ')
-                            print("\ncontinuing...")
-                        else:
-                            both_file = open('Tuple.txt', 'a')
-                            both_file.write('[' + str(avgheight) + ', ' + str(avgmicro) + ']\n')
+                if tracking_objects[i][1] <= 10:
+                    tracking_objects.pop(i)
+                elif last_known != 0 and last_known is not None:
+                    if abs(last_known[0][0] - tracking_objects[i][0][0]) >= 5 or abs(last_known[0][1] - tracking_objects[i][0][1]) >= 5:
+                        tracking_objects.pop(i)
                     else:
-                        if datapointnum == 1:
-                            both_file = open('Tuple.txt', 'a')
-                            both_file.write('[' + str(avgheight) + ', ' + str(avgmicro) + ']\n')
-                        else:
-                            both_file = open('Tuple.txt', 'a')
-                            both_file.write('[' + str(avgheight) + ', ' + str(avgmicro) + ']\n')
-                    heightvec = []
-                    microvec = []
-                datacollect = False
-            else:
+                        x_position = int(tracking_objects[i][0][0])
+                        height = int(tracking_objects[i][1])
+                        y_position = int(tracking_objects[i][0][1])
+                        y_position_adj = (y_end - y_start) - y_position
+            except IndexError:
                 pass
-            if frame_num == total_frames-1:
-                print('Complete')
+
+    return x_position, y_position_adj, height, image_with_keypoints, keypoints_copy
+
+
+def _initialize_tracking(keypoints_cur_frame, keypoints_prev_frame, tracking_objects, track_id):
+    """
+    Initialize tracking for new particles
+    :param keypoints_cur_frame: List object containing tuples of particle locations in the current frame
+    :param keypoints_prev_frame: List object containing tuples of particle locations in the previous frame
+    :param tracking_objects: Dictionary object containing particles' locations
+    :param track_id: Index of particle in tracking_objects dictionary
+    :return track_id: Index of next particle in tracking_objects dictionary
+    """
+    for pt1 in keypoints_cur_frame:
+        for pt2 in keypoints_prev_frame:
+            if math.dist(pt1, pt2) < 10:
+                tracking_objects[track_id] = [pt1]
+                track_id += 1
+    return track_id
+
+
+def _update_tracking(keypoints_cur_frame, tracking_objects):
+    """
+    Update tracking for existing particles in tracking_objects
+    :param keypoints_cur_frame: List object containing tuples of particle locations in the current frame
+    :param tracking_objects: Dictionary object containing particles' locations
+    :return: Updated tracking_objects dictionary for the current frame
+    """
+    tracking_objects_copy = tracking_objects.copy()
+    keypoints_cur_frame_copy = keypoints_cur_frame.copy()
+    
+    for object_id, item2 in tracking_objects_copy.items():
+        object_exists = False
+        for pt1 in keypoints_cur_frame:
+            if math.dist(pt1, item2[0]) < 10:
+                tracking_objects[object_id] = [pt1]
+                object_exists = True
+                if pt1 in keypoints_cur_frame:
+                    keypoints_cur_frame.remove(pt1)
                 break
-        frame_num += 1
+        
+        if not object_exists:
+            tracking_objects.pop(object_id)
 
-    # Initiates data collection if auto is off and we are using the spacebar to manually collect the data points
-    if frames_to_play == sample_frames:
-        if output == "tuple":
-            avgheight = round(np.mean(heightvec), 2)
-            avgmicro = round(np.mean(microvec), 2)
-            print('Average Height = ' + str(avgheight) + ' , Average Micromotion = ' + str(avgmicro))
-            if os.stat('Tuple.txt').st_size == 0:
-                if datapointnum == 1:
-                    input('Tuple.txt already contains data. Press any button to continue to add to it')
-                else:
-                    both_file = open('Tuple.txt', 'a')
-                    both_file.write('[' + str(avgheight) + ', ' + str(avgmicro) + ']\n')
+    track_id_2 = max(tracking_objects.keys(), default=-1) + 1
+    for pt1 in keypoints_cur_frame:
+        tracking_objects[track_id_2] = [pt1]
+        track_id_2 += 1
 
 
-# --------------------------------------- Final Cleanup ---------------------------------------------- #
+def _process_contours(contours, tracking_objects):
+    """
+    Process contours to get particle dimensions
+    :param contours: List object containing contours stored as arrays of points outlining shapes of interest
+    :param tracking_objects: Dictionary object containing particles' locations
+    """
+    for i in contours:
+        x, y, w, h = cv2.boundingRect(i)
+        
+        for key in tracking_objects.keys():
+            if (x <= tracking_objects[key][0][0] <= x + w and 
+                y <= tracking_objects[key][0][1] <= y + h):
+                tracking_objects[key].append(h)
 
 
-# Releasing the VideoCapture object we created
-vid_cap.release()
+def analyze_trial(datapoint):
+    """
+    Analyze trial data and compute averages
+    :param datapoint: List object containing tuples (x,y,h) for particles location and height from each of the previous frames
+    :return: Tuple reflecting the average of the tuples in datapoint
+    """
+    if not datapoint:
+        return 0, 0, 0
+        
+    x = [point[0] for point in datapoint]
+    y = [point[1] for point in datapoint]
+    h = [point[2] for point in datapoint]
+    
+    return (round(np.mean(x), 2),
+            round(np.mean(y), 2),
+            round(np.mean(h), 2))
 
-# Close any open windows
-cv2.destroyAllWindows()
 
+def save_data(yav, hav, frame_num, total_frames, datapoint_num, config = get_default_config()):
+    """
+    Puts height and micromotion data (in millimeters, based on pixel_to_mm parameter) into text file
+    :param yav: Average y-position of the particle over the sample frames, measured from the bottom of the region of interest
+    :param hav: Average height of the particle over the sample frames
+    :param frame_num: Frame number of interest
+    :param total_frames: Total frames contained in the video object
+    :param datapoint_num: Datapoint number, starting at 0
+    :return: Generates or amends the text file in the local directory, places list objects formatted as "[voltage, yav, hav]" on each line
+    """
+    voltage = config.start_voltage + (datapoint_num * config.voltage_increment)
+    cut_file_name = config.video_file.replace('.avi', '')
+    try:
+        if os.stat(str(cut_file_name) + '_data.txt').st_size != 0 and frame_num <= 70:
+            acknowledgement = ""
+            while acknowledgement != "continue":
+                acknowledgement = input(
+                    'File name already contains data. Please cancel and clear the file before proceeding. Type "continue" to override')
+            print("\ncontinuing...")
+        with open(str(cut_file_name) + '_data.txt', 'a') as f:
+            if config.pixel_to_mm == None:
+                conversion = 1
+            else:
+                conversion = config.pixel_to_mm
+            yav_mm = yav * conversion
+            hav_mm = hav * conversion
+            if (yav_mm, hav_mm) != (0, 0):
+                f.write('[' + str(voltage) + ', ' + str(round(yav_mm, 2)) + ', ' + str(round(hav_mm, 2)) + ']\n')
+                percentage = (frame_num / total_frames) * 100
+                print("Saved: " + str(voltage) + ', ' + str(round(yav_mm, 2)) + ', ' + str(round(hav_mm, 2)) + '; Completion: ' + str(round(percentage, 0)) + '%, ' + str(frame_num))
+            else:
+                print('No Particle Detected')
+    except FileNotFoundError:
+        with open(str(cut_file_name) + '_data.txt', 'w') as f:
+            if config.pixel_to_mm == None:
+                conversion = 1
+            else:
+                conversion = config.pixel_to_mm
+            yav_mm = yav * conversion
+            hav_mm = hav * conversion
+            if (yav_mm, hav_mm) != (0, 0):
+                f.write('[' + str(voltage) + ', ' + str(round(yav_mm, 2)) + ', ' + str(round(hav_mm, 2)) + ']\n')
+                percentage = (frame_num / total_frames) * 100
+                print("Saved: " + str(voltage) + ', ' + str(round(yav_mm, 2)) + ', ' + str(round(hav_mm, 2)) + '; Completion: ' + str(round((percentage), 0)) + '%, ' + str(frame_num))
+            else: 
+                print('No Particle Detected')
+
+def auto_run(cap, config = get_default_config()):
+    """
+    Automatic processing of video frames, outputs datapoints as described below in a text data file
+    :param cap: Video capture object from the OpenCV package
+    :return: Generates or amends the text file in the local directory, places list objects formatted as "[voltage, y-position, height]" on each line
+    """
+    total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+    tracking_objects, track_id, keypoints_prev_frame = setup_tracking()
+    _, _, y_start, y_end = gen_initial_frame(cap)
+    collection_frames = [
+        int((config.fps * config.change_interval * i) + 
+            (config.fps * config.change_interval * 0.4))
+        for i in range(100)
+    ]
+    end_collection_frames = [cf + config.sample_frames for cf in collection_frames]
+    _, datapoint = [], []
+    collect_data = False
+    keypoints_prev_frame = []
+    datapoint_num = 0
+    for frame_num in range(config.start_frame, total_frames):
+        ret, frame = get_frame(cap, frame_num)
+        if frame_num >= config.start_frame + 2 and len(tracking_objects.keys()) > 0 :
+            for i in tracking_objects.keys():
+                try:
+                    last_known = tracking_objects[i]
+                except KeyError:
+                    pass
+        tracking_objects, track_id, _ = setup_tracking()
+        if not ret:
+            break
+        roi_frame, closing, _, _ = post_processing(cap, frame, frame_num)
+        if frame_num >= config.start_frame + 2:
+            x, y, h, _, keypoints_cur_frame = locate_particles(roi_frame, closing, keypoints_prev_frame, 
+                                 frame_num, tracking_objects, track_id, y_end, y_start, last_known)
+        else:
+            x, y, h, _, keypoints_cur_frame = locate_particles(roi_frame, closing, keypoints_prev_frame, 
+                                 frame_num, tracking_objects, track_id, y_end, y_start)
+        keypoints_prev_frame = keypoints_cur_frame
+        if frame_num in collection_frames:
+            collect_data = True
+        if frame_num in end_collection_frames:
+            collect_data = False
+            xav, yav, hav = analyze_trial(datapoint)
+            save_data(yav, hav, frame_num, total_frames, datapoint_num)
+            datapoint_num = datapoint_num + 1
+            datapoint = []
+        if collect_data and x != "NaN":
+            datapoint.append([x, y, h])
+
+
+def run_frame(cap, frame_num, keypoints_prev_frame):
+    """
+    Manually processes and displays each frame. Press a letter or arrow key to progress
+    :param cap: Video capture object from the OpenCV package
+    :param frame_num: Frame number of interest
+    :param keypoints_prev_frame: List object containing tuples of particle locations in the previous frame
+    :return frame_num: Frame number of the next frame to analyze
+    :return keypoints_cur_frame: List object containing detected objects of current frame
+    """
+    tracking_objects, track_id, _ = setup_tracking()
+    ret, frame = get_frame(cap, frame_num)
+    if not ret:
+        exit()
+    _, _, y_start, y_end = gen_initial_frame(cap)
+    roi_frame, closing, _, closing_raw = post_processing(cap, frame, frame_num)
+    _, _, _, image_with_keypoints, keypoints_cur_frame = locate_particles(roi_frame, closing, keypoints_prev_frame, 
+                                frame_num, tracking_objects, track_id, y_end, y_start)
+    
+    cv2.imshow("Frame", closing_raw)
+
+    frame_num = frame_num + 1
+    return frame_num, keypoints_cur_frame
+        
+
+def main():
+    """
+    Main entry point
+    """
+    
+    config = TrackingConfig()
+
+    cap = cv2.VideoCapture(config.video_file)
+    _, _, _, _ = gen_initial_frame(cap)
+    
+    frame_num = config.start_frame
+    for i in range(int(cap.get(cv2.CAP_PROP_FRAME_COUNT))):
+        if i == 0:
+            keypoints_prev_frame = []
+        frame_num, keypoints_prev_frame = run_frame(cap, frame_num, keypoints_prev_frame)
+        key = cv2.waitKey()
+        if key == 27:  # ESC
+            exit()
+        if key == 32:  # Space
+            auto_run(cap)
+        else:
+            pass
+
+if __name__ == "__main__":
+    main()
